Split routes in two
So, to get a direct link it's better to have one route that calls a
mutation as soon as it is visited.
