+++ webapp/locales/en.json
@@ -166,7 +166,9 @@
       "labelNonce": "Enter your code",
       "success": "A new E-Mail address has been registered.",
       "submitted": "An email to verify your address has been sent to <b>{email}</b>.",
-      "change-successful": "Your E-Mail address has been changed successfully."
+      "change-successful": "Your E-Mail address has been changed successfully.",
+      "change-error": "Your E-Mail could not be changed.",
+      "change-error-help": "Maybe the code was invalid or you did not add a new E-Mail address before?"
     },
     "validation": {
       "slug": {
